feat(utilities): send email to the customer
