+++ package.json
@@ -11,6 +11,7 @@
     "express": "~4.16.0",
     "moment": "^2.22.2",
     "mongodb": "^3.1.4",
-    "morgan": "~1.9.0"
+    "morgan": "~1.9.0",
+    "nodemailer": "^4.6.8"
   }
 }
